[refactor] 각 player 별로 frameSet 을 관리하도록 변경
